Generalized verticals error presentation for reuse with Domains
