+++ WordPress/Classes/ViewRelated/Site Creation/Shared/TableDataCoordinator.swift
@@ -1,6 +1,9 @@
 import UIKit
 
+typealias TableViewProvider = UITableViewDataSource & UITableViewDelegate
+
 /// Generic-based implementation of the UITableViewDataSource and UITableViewDelegate protocol.
+///
 final class TableDataCoordinator<Model, Cell>: NSObject, UITableViewDataSource, UITableViewDelegate where Cell: ModelSettableCell, Cell: UITableViewCell, Model == Cell.DataType {
     private let data: [Model]
     private let selection: (Model) -> Void
